requirements.txt
setuptools==67.1.0
dj-database-url==0.5.0
Django==4.1.6
PyYAML==5.1
djangorestframework
uritemplate
Markdown
django-cors-headers
django-environ==0.4.5
django-extensions==3.1.2
django-filter==21.1
django-simple-history==2.12.0
django-tenant-users==1.1.1
django-tenants==3.4.7
drf-writable-nested==0.6.3
gunicorn==20.1.0
psycopg2-binary==2.9.1
Pygments==2.2.0
PyNaCl==1.4.0
pyparsing==3.0.7
python-dateutil==2.8.2
python-decouple==3.4
drf-yasg
coreapi
django-rest-swagger
openapi-codec

